fix : 검색결과 없는 경우 tag_id 204

+++ server/routers/hashtag.py
@@ -149,31 +149,31 @@ async def post_hashtags(hastag : Hashtag):
 
 
     #조인 테이블 생성
-    await tags_union(tag_id, hastag.hashtags)
-
-    
-    if(hastag.isAds): #광고제거
-        await db[tag_id].delete_many({"isAds": True})
-    if(hastag.period>0): #기간필터
-        current_date = datetime.now()
-        days_ago = current_date - timedelta(days=hastag.period)
-        print(days_ago)
-        await db[tag_id].delete_many({"date": {
-            "$lte": days_ago
-        }})
-    
-    #AI 모델 
-    if hastag.image_url is not None :
-        cursor = db[tag_id].find()
-        docs = await cursor.to_list(length=50)
-        img_docs = list()
-        for doc in docs:
-            if doc["image_url"] is not None:
-                img_docs.append({"id":str(doc["_id"]), "image_url":doc["image_url"]})
-        sort_images = model_predict(hastag.image_url, img_docs)
-        for idx, sort in enumerate(sort_images):
-            await db[tag_id].find_one_and_update({"_id":ObjectId(sort["id"])},{"$set":{"image_rank":idx+1}})
-        await db[tag_id].delete_many({"image_rank":0})
+    if(await tags_union(tag_id, hastag.hashtags)):
+        if(hastag.isAds): #광고제거
+            await db[tag_id].delete_many({"isAds": True})
+        if(hastag.period>0): #기간필터
+            current_date = datetime.now()
+            days_ago = current_date - timedelta(days=hastag.period)
+            print(days_ago)
+            await db[tag_id].delete_many({"date": {
+                "$lte": days_ago
+            }})
+        
+        #AI 모델 
+        if hastag.image_url is not None :
+            cursor = db[tag_id].find()
+            docs = await cursor.to_list(length=50)
+            img_docs = list()
+            for doc in docs:
+                if doc["image_url"] is not None:
+                    img_docs.append({"id":str(doc["_id"]), "image_url":doc["image_url"]})
+            sort_images = model_predict(hastag.image_url, img_docs)
+            for idx, sort in enumerate(sort_images):
+                await db[tag_id].find_one_and_update({"_id":ObjectId(sort["id"])},{"$set":{"image_rank":idx+1}})
+            await db[tag_id].delete_many({"image_rank":0})
+    else: 
+        await db["tagsId"].find_one_and_delete({"_id" : ObjectId(tag_id)})
     
     print("total time// amount:"+str(amount)+" timetotal:", time.time() - start)
     return {"status": 201, "message": "검색 종료"}
@@ -238,7 +238,10 @@ async def tags_union(tag_id:str, hashtags : List[str] = Query(None)):
             db[tag_id].delete_many({"_id": {"$in": duplicate_ids}})
     
         
-    await update_sort(tag_id, True, False, False)
+        await update_sort(tag_id, True, False, False)
+        return True
+    else:
+        return False
  
  #
  ##[GET] 해시태그 검색 결과
